requirements.txt
# Core requirements
Django>=4.2.0
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0
nltk>=3.6.0
python-dotenv>=0.19.0
joblib>=1.1.0
celery>=5.3.0
django-compressor>=4.1
gunicorn>=20.1.0
whitenoise>=6.4.0
redis>=4.5.0
django-redis>=5.2.0
django-debug-toolbar>=4.0.0

# Testing and development
pytest>=7.3.1
pytest-django>=4.5.2
coverage>=7.2.3
black>=23.3.0
flake8>=6.0.0

# Configuration and utilities
python-decouple>=3.8
Pillow>=10.0.0
piexif>=1.1.3
gevent>=23.7.0

# Production database (optional)
# psycopg2-binary>=2.9.6  # Uncomment for PostgreSQL support